requirements.txt
streamlit
torch
torchvision
ultralytics
pandas
Pillow